chore(bpmn): rename bpmn modeler
