+++ client/src/app/tabs/bpmn/modeler/BpmnModeler.js
@@ -3,21 +3,18 @@ import BpmnModeler from 'bpmn-js/lib/Modeler';
 import diagramOriginModule from 'diagram-js-origin';
 import minimapModule from 'diagram-js-minimap';
 
-import addExporterModule from '@bpmn-io/add-exporter';
-
 import executableFixModule from './features/executable-fix';
 import globalClipboardModule from './features/global-clipboard';
-import applyDefaultTemplates from './features/apply-default-templates';
 import propertiesPanelKeyboardBindingsModule from './features/properties-panel-keyboard-bindings';
 
 import signavioCompatModule from 'bpmn-js-signavio-compat';
 
-import camundaModdlePackage from 'camunda-bpmn-moddle/resources/camunda';
-import camundaModdleExtension from 'camunda-bpmn-moddle/lib';
-
 import propertiesPanelModule from 'bpmn-js-properties-panel';
-import propertiesProviderModule from 'bpmn-js-properties-panel/lib/provider/camunda';
+import propertiesProviderModule from '../custom/properties-provider';
 
+import zeebeModdleExtension from '../custom/zeebe-bpmn-moddle/zeebe';
+
+import zeebeCustoms from '../custom';
 
 import 'bpmn-js-properties-panel/styles/properties.less';
 
@@ -27,7 +24,7 @@ import 'bpmn-js/dist/assets/bpmn-font/css/bpmn-embedded.css';
 import 'diagram-js-minimap/assets/diagram-js-minimap.css';
 
 
-export default class CamundaBpmnModeler extends BpmnModeler {
+export default class ZeebeBpmnModeler extends BpmnModeler {
 
   constructor(options = {}) {
 
@@ -39,7 +36,7 @@ export default class CamundaBpmnModeler extends BpmnModeler {
     super({
       ...otherOptions,
       moddleExtensions: {
-        camunda: camundaModdlePackage,
+        zeebe: zeebeModdleExtension,
         ...(moddleExtensions || {})
       }
     });
@@ -51,18 +48,17 @@ const defaultModules = BpmnModeler.prototype._modules;
 const extensionModules = [
   diagramOriginModule,
   minimapModule,
-  addExporterModule,
   executableFixModule,
   globalClipboardModule,
   signavioCompatModule,
-  camundaModdleExtension,
   propertiesPanelModule,
   propertiesProviderModule,
-  applyDefaultTemplates,
-  propertiesPanelKeyboardBindingsModule
+  propertiesPanelKeyboardBindingsModule,
+  zeebeModdleExtension,
+  zeebeCustoms
 ];
 
-CamundaBpmnModeler.prototype._modules = [
+ZeebeBpmnModeler.prototype._modules = [
   ...defaultModules,
   ...extensionModules
 ];
